+++ lib/cli/cli.ex
@@ -15,6 +15,7 @@ defmodule Doctor.CLI do
     Project.config()
     |> Keyword.get(:app)
     |> get_application_modules()
+    |> Enum.filter(fn module -> String.starts_with?(to_string(module), "Elixir.") end)
 
     # Fetch the module information from the list of application modules
     |> Enum.map(&generate_module_entry/1)
@@ -39,20 +40,18 @@ defmodule Doctor.CLI do
     Project.config()
     |> Keyword.get(:app)
     |> get_application_modules()
-    |> Enum.find(:not_found, fn module ->
-      Atom.to_string(module) == "Elixir.#{module_name}"
-    end)
+    |> Enum.find(:not_found, &(inspect(&1) == module_name))
     |> case do
       :not_found ->
-        raise "Could not find module #{inspect(module_name)} in application"
+        :not_found
 
       module ->
         module
         |> generate_module_entry()
         |> async_fetch_user_defined_functions()
         |> Task.await(15_000)
+        |> ModuleExplain.generate_report(args)
     end
-    |> ModuleExplain.generate_report(args)
   end
 
   @doc """